# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Investintell Light API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/portfolios": {
      "get": {
        "tags": [
          "portfolios"
        ],
        "summary": "List Portfolios",
        "description": "List portfolios (id order), hard-capped at the service's LIST_HARD_CAP.",
        "operationId": "list_portfolios_portfolios_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PortfolioListItem"
                  },
                  "type": "array",
                  "title": "Response List Portfolios Portfolios Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "portfolios"
        ],
        "summary": "Create Portfolio",
        "description": "Create a portfolio with optional initial positions.\n\nPosition tickers are validated against Tiingo via the EOD ensure \u2014 a typo\nfails loud (404) BEFORE anything is persisted \u2014 which also warms the EOD\ncache for the overview.",
        "operationId": "create_portfolio_portfolios_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PortfolioCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PortfolioOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/portfolios/{portfolio_id}": {
      "get": {
        "tags": [
          "portfolios"
        ],
        "summary": "Get Portfolio",
        "description": "One portfolio with its positions (sorted by ticker).",
        "operationId": "get_portfolio_portfolios__portfolio_id__get",
        "parameters": [
          {
            "name": "portfolio_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Portfolio Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PortfolioOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "portfolios"
        ],
        "summary": "Patch Portfolio",
        "description": "Partially update name and/or cash.",
        "operationId": "patch_portfolio_portfolios__portfolio_id__patch",
        "parameters": [
          {
            "name": "portfolio_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Portfolio Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PortfolioPatch"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PortfolioOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "portfolios"
        ],
        "summary": "Delete Portfolio",
        "description": "Delete a portfolio; its positions cascade away at the DB level.",
        "operationId": "delete_portfolio_portfolios__portfolio_id__delete",
        "parameters": [
          {
            "name": "portfolio_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Portfolio Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/portfolios/{portfolio_id}/positions/{ticker}": {
      "put": {
        "tags": [
          "portfolios"
        ],
        "summary": "Put Position",
        "description": "Upsert one position.\n\nINSERT path validates the ticker against Tiingo (and warms the EOD cache);\nthe UPDATE path deliberately does NOT re-ensure \u2014 the ticker was already\nvalidated when the position was created.\n\nF8.6b: the body optionally carries basis/commission/trade_date (manual\nfill registration). Fields absent from the body keep the stored values\non UPDATE; on INSERT basis defaults to 'reference'.",
        "operationId": "put_position_portfolios__portfolio_id__positions__ticker__put",
        "parameters": [
          {
            "name": "portfolio_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Portfolio Id"
            }
          },
          {
            "name": "ticker",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Ticker"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PositionBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PositionOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "portfolios"
        ],
        "summary": "Delete Position",
        "description": "Delete one position; 404 covers both a missing portfolio and a missing ticker.",
        "operationId": "delete_position_portfolios__portfolio_id__positions__ticker__delete",
        "parameters": [
          {
            "name": "portfolio_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Portfolio Id"
            }
          },
          {
            "name": "ticker",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Ticker"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/portfolios/{portfolio_id}/overview": {
      "get": {
        "tags": [
          "portfolios"
        ],
        "summary": "Get Portfolio Overview",
        "description": "Render-ready position table with P&L and column-header aggregates (D6).\n\nLast/prev closes come from the two most recent eod_prices rows per ticker;\nthe EOD ensure runs first so stale tickers are refreshed.  An empty\nportfolio is a legitimate 200 with zeroed/null aggregates.",
        "operationId": "get_portfolio_overview_portfolios__portfolio_id__overview_get",
        "parameters": [
          {
            "name": "portfolio_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Portfolio Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PortfolioOverviewResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/portfolios/{portfolio_id}/news": {
      "get": {
        "tags": [
          "portfolios"
        ],
        "summary": "Get Portfolio News",
        "description": "Aggregated news across all portfolio tickers, newest first.\n\nStaleness is checked per ticker but all stale tickers are refreshed with\nONE combined Tiingo call (see ``ensure_news``).  Degrade path mirrors\nGET /stocks/{ticker}/news exactly: refresh failure with cached articles\nserves them with ``stale=true``; with an empty cache it fails loud.",
        "operationId": "get_portfolio_news_portfolios__portfolio_id__news_get",
        "parameters": [
          {
            "name": "portfolio_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Portfolio Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 50,
              "minimum": 1,
              "description": "Max articles returned.",
              "default": 20,
              "title": "Limit"
            },
            "description": "Max articles returned."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PortfolioNewsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PortfolioCreate": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name",
            "description": "Portfolio name; 1..80 characters after trimming, unique across the installation."
          },
          "cash": {
            "type": "number",
            "title": "Cash",
            "description": "Uninvested cash, currency units.",
            "default": 0.0
          },
          "positions": {
            "items": {
              "$ref": "#/components/schemas/PositionCreate"
            },
            "type": "array",
            "maxItems": 50,
            "title": "Positions",
            "description": "Initial positions (at most 50); tickers must be unique."
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "PortfolioCreate",
        "description": "Body for POST /portfolios."
      },
      "PortfolioListItem": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "cash": {
            "type": "number",
            "title": "Cash"
          },
          "position_count": {
            "type": "integer",
            "title": "Position Count"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "cash",
          "position_count",
          "created_at"
        ],
        "title": "PortfolioListItem",
        "description": "One row of GET /portfolios (positions omitted, count only)."
      },
...
      "PortfolioOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "cash": {
            "type": "number",
            "title": "Cash"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "positions": {
            "items": {
              "$ref": "#/components/schemas/PositionOut"
            },
            "type": "array",
            "title": "Positions"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "cash",
          "created_at",
          "updated_at",
          "positions"
        ],
        "title": "PortfolioOut",
        "description": "One persisted portfolio with its positions (sorted by ticker)."
      },
      "PortfolioOverviewResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "positions": {
            "items": {
              "$ref": "#/components/schemas/PositionOverview"
            },
            "type": "array",
            "title": "Positions"
          },
          "aggregates": {
            "$ref": "#/components/schemas/OverviewAggregates"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "positions",
          "aggregates"
        ],
        "title": "PortfolioOverviewResponse",
        "description": "Render-ready overview for GET /portfolios/{id}/overview."
      },
...
      "PortfolioPatch": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name",
            "description": "New portfolio name (same rules as on create)."
          },
          "cash": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cash",
            "description": "New cash balance, currency units."
          }
        },
        "type": "object",
        "title": "PortfolioPatch",
        "description": "Body for PATCH /portfolios/{id} \u2014 at least one field must be present."
      },
...
      "PositionBody": {
        "properties": {
          "quantity": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Quantity",
            "description": "Shares/units held; > 0."
          },
          "acq_price": {
            "anyOf": [
              {
                "type": "number",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Acq Price",
            "description": "Acquisition price per share/unit; null = unknown (P&L renders null). With basis='executed' this is the effective cost basis incl. commissions."
          },
          "basis": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "reference",
                  "executed"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Basis",
            "description": "'reference' (spot/NAV for analysis) or 'executed' (real fill incl. commissions). Omitted: insert defaults to 'reference', update keeps the stored value."
          },
          "commission": {
            "anyOf": [
              {
                "type": "number",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Commission",
            "description": "Total commission paid on the fill, currency units (>= 0)."
          },
          "trade_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trade Date",
            "description": "Execution date of the fill."
          }
        },
        "type": "object",
        "required": [
          "quantity"
        ],
        "title": "PositionBody",
        "description": "Quantity/acquisition-price payload for the position upsert (PUT).\n\nF8.6b fill fields are optional: omitting them keeps the pre-F8.6b\nbehavior (on UPDATE the stored basis/commission/trade_date are left\nuntouched; on INSERT basis defaults to 'reference')."
      },
...
      "PositionOut": {
        "properties": {
          "ticker": {
            "type": "string",
            "title": "Ticker"
          },
          "quantity": {
            "type": "number",
            "title": "Quantity"
          },
          "acq_price": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Acq Price"
          },
          "basis": {
            "type": "string",
            "enum": [
              "reference",
              "executed"
            ],
            "title": "Basis"
          },
          "commission": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Commission"
          },
          "trade_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trade Date"
          }
        },
        "type": "object",
        "required": [
          "ticker",
          "quantity",
          "acq_price",
          "basis",
          "commission",
          "trade_date"
        ],
        "title": "PositionOut",
        "description": "One persisted position."
      },
...
    }
  }
}